Feature API Courses (#2)
Added APIs with the help of jsonapi-resources.
Added pagination as well with the help of jsonapi-resources.
Added test coverage with the help of rspec and make sure things are working as expected.
There was a use case where A coach can be removed/deleted, transfer the course to another coach in case the deleted
coach is responsible for a course. I did make sure with test cases that we are transferring course to another coach.
In case we do not have another coach to assign the course we prevent deletion of coach.
Other use cases are pretty simple and I have added Api requests in README as well.